Fix double headers
